Rename A22 -> A23
Since A22 should now be the unit conversion, the old function for calculating parameters has been renamed A23.
