+++ bin/A23.py
