+++ bin/__init__.py
@@ -1,7 +1,7 @@
 #Initialize the package directory using this python file
 #from .A11 import A11
 from .A21 import A21
-from .A22 import A22
+from .A23 import A23
 from .A31 import A31
 #from .A32 import A32
 from .suddenContractionKtRegression import contractKt
